+++ NEWS.md
@@ -1,4 +1,4 @@
-# AMR 0.7.1.9006
+# AMR 0.7.1.9007
 
 ### New
 * Additional way to calculate co-resistance, i.e. when using multiple antibiotics as input for `portion_*` functions or `count_*` functions. This can be used to determine the empiric susceptibily of a combination therapy. A new parameter `only_all_tested` (**which defaults to `FALSE`**) replaces the old `also_single_tested` and can be used to select one of the two methods to count isolates and calculate portions. The difference can be seen in this example table (which is also on the `portion` and `count` help pages), where the %SI is being determined:
@@ -37,6 +37,8 @@
 * Fix and speed improvement for `mo_shortname()`
 * Fix for `as.mo()` where misspelled input would not be understood
 * Fix for using `mo_*` functions where the coercion uncertainties and failures would not be available through `mo_uncertainties()` and `mo_failures()` anymore
+* Deprecated the `country` parameter of `mdro()` in favour of the already existing `guideline` parameter to support multiple guidelines within one country
+* Fix for frequency tables when creating one directly on a group (using `group_by()`)
 
 # AMR 0.7.1